pyproject.toml
[tool.poetry]
name = "llm-mlops"
version = "0.1.0"
description = ""
authors = ["Arvin Bk <arvin.balak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
transformers = "^4.39.3"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
